refactor: 기존 getAccommodationScoreMap, getRoomScoreMap, calculateScore, calculateScoreAndSort 메서드들을 PetFacilityScoreService 기반으로 수정
- AccommodationRecommendationService를 recommendation 패키지로 이동
